requirements.txt
apscheduler==3.10.4
pymongo==4.6.0
python-dotenv==1.0.0
Pillow==10.2.0
opencv-python==4.8.1.78
pymediainfo==6.1.0
pandas==2.1.4
pydantic==2.5.0
pydantic-settings==2.1.0
pyodbc==5.3.0
pytz==2023.3
pyyaml==6.0.3
streamlit==1.28.1
plotly==5.18.0
